requirements/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
asgiref==3.11.1
cffi==2.0.0
dj-config-url==0.1.1
dj-database-url==3.1.2
Django==6.0.3
django-axes==8.3.1
django-cors-headers==4.9.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework-stubs==3.16.8
djangorestframework_simplejwt==5.5.1
gunicorn==25.1.0
packaging==26.0
pillow==12.1.1
pycparser==3.0
PyJWT==2.11.0
python-decouple==3.8
python-magic==0.4.27
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
whitenoise==6.12.0
